refactor(apps/twitch-bot): rename the PostgreSQL package to MySQL
